+++ src/requests/mod.rs
@@ -9,6 +9,7 @@ mod set_device_info;
 mod tapo_request;
 
 pub use color::*;
+pub use set_device_info::*;
 
 pub(crate) use get_device_info::*;
 pub(crate) use get_device_usage::*;
@@ -16,5 +17,4 @@ pub(crate) use get_energy_usage::*;
 pub(crate) use handshake::*;
 pub(crate) use login_device::*;
 pub(crate) use secure_passthrough::*;
-pub(crate) use set_device_info::*;
 pub(crate) use tapo_request::*;