+++ src/requests/set_device_info/mod.rs
@@ -2,6 +2,7 @@ mod generic_device;
 mod l510;
 mod l530;
 
+pub use l510::*;
+pub use l530::*;
+
 pub(crate) use generic_device::*;
-pub(crate) use l510::*;
-pub(crate) use l530::*;